+++ mkkey/cli.py
@@ -49,16 +49,16 @@ def _jwk(
     return
 
 
-def _paserk_public(version: int, kid: bool, password: str, rsa_key_size: int = 2048):
+def _paserk_public(version: int, kid: bool, password: str, wrapping_key: str, rsa_key_size: int = 2048):
     try:
-        _show_result(generate_public_paserk(version, kid, password, rsa_key_size=rsa_key_size))
+        _show_result(generate_public_paserk(version, kid, password, wrapping_key, rsa_key_size=rsa_key_size))
     except Exception as err:
         _show_error(err)
 
 
-def _paserk_local(version: int, key_material: str, kid: bool, password: str):
+def _paserk_local(version: int, key_material: str, kid: bool, password: str, wrapping_key: str = ""):
     try:
-        _show_result(generate_local_paserk(version, key_material, kid, password))
+        _show_result(generate_local_paserk(version, key_material, kid, password, wrapping_key))
     except Exception as err:
         _show_error(err)
 
@@ -360,9 +360,16 @@ def v4():
     required=False,
     help="Set password for key wrapping.",
 )
-def paserk_v4_public(kid: bool, password: str):
+@click.option(
+    "--wrapping-key",
+    type=str,
+    default="",
+    required=False,
+    help="Set another symmetric key for key wrapping.",
+)
+def paserk_v4_public(kid: bool, password: str, wrapping_key: str):
     """Generate v4.public PASERK for Asymmetric-key digital signatures."""
-    _paserk_public(4, kid, password)
+    _paserk_public(4, kid, password, wrapping_key)
     return
 
 
@@ -386,7 +393,14 @@ def paserk_v4_public(kid: bool, password: str):
     required=False,
     help="Set password for key wrapping.",
 )
-def paserk_v4_local(key_material: str, kid: bool, password: str):
+@click.option(
+    "--wrapping-key",
+    type=str,
+    default="",
+    required=False,
+    help="Set another symmetric key for key wrapping.",
+)
+def paserk_v4_local(key_material: str, kid: bool, password: str, wrapping_key: str):
     """Generate v4.local PASERK for Symmetric-key encryption (AEAD)."""
     _paserk_local(4, key_material, kid, password)
     return
@@ -411,9 +425,16 @@ def v3():
     required=False,
     help="Set password for key wrapping.",
 )
-def paserk_v3_public(kid: bool, password: str):
+@click.option(
+    "--wrapping-key",
+    type=str,
+    default="",
+    required=False,
+    help="Set another symmetric key for key wrapping.",
+)
+def paserk_v3_public(kid: bool, password: str, wrapping_key: str):
     """Generate v3.public PASERK for Asymmetric-key digital signatures."""
-    _paserk_public(3, kid, password)
+    _paserk_public(3, kid, password, wrapping_key)
     return
 
 
@@ -437,9 +458,16 @@ def paserk_v3_public(kid: bool, password: str):
     required=False,
     help="Set password for key wrapping.",
 )
-def paserk_v3_local(key_material: str, kid: bool, password: str):
+@click.option(
+    "--wrapping-key",
+    type=str,
+    default="",
+    required=False,
+    help="Set another symmetric key for key wrapping.",
+)
+def paserk_v3_local(key_material: str, kid: bool, password: str, wrapping_key: str):
     """Generate v3.local PASERK for Symmetric-key encryption (AEAD)."""
-    _paserk_local(3, key_material, kid, password)
+    _paserk_local(3, key_material, kid, password, wrapping_key)
     return
 
 
@@ -462,9 +490,16 @@ def v2():
     required=False,
     help="Set password for key wrapping.",
 )
-def paserk_v2_public(kid: bool, password: str):
+@click.option(
+    "--wrapping-key",
+    type=str,
+    default="",
+    required=False,
+    help="Set another symmetric key for key wrapping.",
+)
+def paserk_v2_public(kid: bool, password: str, wrapping_key: str):
     """Generate v2.public PASERK for Asymmetric-key digital signatures."""
-    _paserk_public(2, kid, password)
+    _paserk_public(2, kid, password, wrapping_key)
     return
 
 
@@ -488,9 +523,16 @@ def paserk_v2_public(kid: bool, password: str):
     required=False,
     help="Set password for key wrapping.",
 )
-def paserk_v2_local(key_material: str, kid: bool, password: str):
+@click.option(
+    "--wrapping-key",
+    type=str,
+    default="",
+    required=False,
+    help="Set another symmetric key for key wrapping.",
+)
+def paserk_v2_local(key_material: str, kid: bool, password: str, wrapping_key: str):
     """Generate v2.local PASERK for Symmetric-key encryption (AEAD)."""
-    _paserk_local(2, key_material, kid, password)
+    _paserk_local(2, key_material, kid, password, wrapping_key)
     return
 
 
@@ -514,6 +556,13 @@ def v1():
     help="Set password for key wrapping.",
 )
 @click.option(
+    "--wrapping-key",
+    type=str,
+    default="",
+    required=False,
+    help="Set another symmetric key for key wrapping.",
+)
+@click.option(
     "--key-size",
     type=int,
     default=2048,
@@ -521,9 +570,9 @@ def v1():
     required=False,
     help="Set the length of modulus in bits for RSA key (MUST be >=512).",
 )
-def paserk_v1(kid: bool, password: str, key_size: int):
+def paserk_v1_public(kid: bool, password: str, wrapping_key: str, key_size: int):
     """Generate v1.public PASERK for Asymmetric-key digital signatures."""
-    _paserk_public(1, kid, password, rsa_key_size=key_size)
+    _paserk_public(1, kid, password, wrapping_key, rsa_key_size=key_size)
     return
 
 
@@ -547,7 +596,14 @@ def paserk_v1(kid: bool, password: str, key_size: int):
     required=False,
     help="Set password for key wrapping.",
 )
-def paserk_v1_local(key_material: str, kid: bool, password: str):
+@click.option(
+    "--wrapping-key",
+    type=str,
+    default="",
+    required=False,
+    help="Set another symmetric key for key wrapping.",
+)
+def paserk_v1_local(key_material: str, kid: bool, password: str, wrapping_key: str):
     """Generate v1.local PASERK for Symmetric-key encryption (AEAD)."""
-    _paserk_local(1, key_material, kid, password)
+    _paserk_local(1, key_material, kid, password, wrapping_key)
     return